cmake: Use correct link option for gcc ld in MinGW.
While MSVC needs to specify entry as `wWinMainCRTStartup`, MinGW
needs `-municode`. MinGW used to not support `wWinMain`, but recently it
finally adds support for that.

The support for clang is more complicated. This code was tested on MinGW
clang, provided by MSYS2 clang64, and it worked successfully. Further
investigation may be needed.

Signed-off-by: Yuyi Wang <Strawberry_Str@hotmail.com>